Added support for the deterministic fault injection. It is now possible to choose the exact time step at which a fault will be injected. Added support for a base Workspace variables as input in the GUI parameters.

+++ Fault_injection/es_inject_error_bitflip.m
@@ -28,7 +28,7 @@ function error_data = es_inject_error_bitflip( input_data, number_of_bitflips )
         res_data = typecast(data_cast, var_type);
     end
 
-    error_data = res_data;
+    error_data = double(res_data);
 
     
 end
